feat: consumable category
